requirements.txt
dnspython==2.7.0
Flask==3.0.3
gunicorn
gevent
python-dotenv==1.0.1
